+++ Domain/Dtos/EditSubscriptionDto.cs
@@ -6,6 +6,6 @@ public class EditSubscriptionDto
     public string? NewName { get; set; }
     public string? NewDescription { get; set; }
     public int? NewMaxResolution { get; set; }
-    public List<int>? AccessibleContentIdsToAdd { get; set; }
-    public List<int>? AccessibleContentIdsToRemove { get; set; }
+    public List<long>? AccessibleContentIdsToAdd { get; set; }
+    public List<long>? AccessibleContentIdsToRemove { get; set; }
 }
