+++ Domain/Dtos/NewSubscriptionDto.cs
@@ -5,5 +5,5 @@ public class NewSubscriptionDto
     public required string Name { get; set; }
     public required string Description { get; set; }
     public int MaxResolution { get; set; } = 720;
-    public List<int> AccessibleContentIds { get; set; } = [];
+    public List<long> AccessibleContentIds { get; set; } = [];
 }
